requirements.txt
django
django-ckeditor 
pillow